New module TokenManagement.scala
changelog_begin
changelog_end
